Decompose DynamicReportDetailsPage into leaf components
